[fix] Post member find 수정 (#140)
* [fix] 게시판 request 수정

[fix] Post member find 수정
what : 게시판 글 등록시 멤버 데이터를 찾을시 함수 변경
why : findByEmail -> findById가 더 적합

* [fix] 게시판 delete 수정

[fix] 게시판 delete 수정
what : 게시판 글 삭제시 멤버 판별코드 추가
why : 게시판 글 삭제시 멤버 판별코드가 존재 안했음

+++ Backend/src/main/java/com/github/backend/service/post/impl/PostServiceImpl.java
@@ -17,6 +17,8 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Objects;
+
 @RequiredArgsConstructor
 @Service
 public class PostServiceImpl implements PostService {
@@ -35,7 +37,7 @@ public class PostServiceImpl implements PostService {
                 .orElseThrow(() -> new PostException(PostErrorCode.POST_NOT_EXISTS));
 
 
-        if (post.getMember().getId() != memberInfo.getId()) {
+        if (!Objects.equals(post.getMember().getId(), memberInfo.getId())) {
             throw new PostException(PostErrorCode.POST_NOT_EQ_MEMBER);
         }
 
@@ -47,7 +49,7 @@ public class PostServiceImpl implements PostService {
     @Transactional
     @Override
     public PostRegisterOutPutDto.Info registerPost(PostReq.Request req, MemberInfo memberInfo) {
-        Member member = memberRepository.findByEmail(memberInfo.getEmail())
+        Member member = memberRepository.findById(memberInfo.getId())
                 .orElseThrow(() -> new PostException(PostErrorCode.MEMBER_NOT_EXISTS));
 
         Post post = postRepository.save(Post.builder()
@@ -61,9 +63,17 @@ public class PostServiceImpl implements PostService {
     }
 
     @Override
-    public PostInfoOutPutDto.Info deletePost(Long postId) {
+    public PostInfoOutPutDto.Info deletePost(Long postId, MemberInfo memberInfo) {
+        Member member = memberRepository.findById(memberInfo.getId())
+                .orElseThrow(() -> new PostException(PostErrorCode.MEMBER_NOT_EXISTS));
+
         Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new PostException(PostErrorCode.POST_NOT_EXISTS));
+
+        if(!Objects.equals(member.getId(), post.getMember().getId())){
+            throw new PostException(PostErrorCode.MEMBER_NOT_EXISTS);
+        }
+
         postRepository.delete(post);
 
         return PostInfoOutPutDto.Info.of((post));
